Fix non-functionality on Windows.
Because of the .exe extension of Go-produced binaries on Windows,
binstale could not tell that those binaries came from similarly-named Go
commands.

Arguments to binstale are expected to be command names, so they do not
include ".exe" suffix even on Windows. However, ".exe" suffix will be
printed when referencing files that are not found.

Fixes #4.
Closes #3.