Display stale reason for stale binaries.
Go 1.7 adds support to go list command to display the reason a binary
is considered stale (i.e., meaning that installing it would have a
non-zero effect).

Reference: https://golang.org/doc/go1.7#cmd_go

We can capture and display that reason, when displaying that a binary
is stale. This extra information may be helpful in understanding why a
binary is considered stale.

Update Travis to Go 1.7.

Update example output in documentation to include stale reason, and
regenerate README.